apps/manga/requirements.txt
Flask==3.0.3
Pillow==10.4.0
anthropic==0.40.0
Werkzeug==3.0.3
gunicorn==23.0.0
openai>=1.0.0
numpy>=1.20.0
moviepy>=1.0.3
edge-tts>=6.1.0